remove v1 of thoreau, and leave just v2
