+++ .ruby-version
@@ -1,1 +1,1 @@
-2.6.3
+2.7.0
