+++ lib/thoreau.rb
@@ -1,8 +1,4 @@
 require_relative 'thoreau/version'
-require_relative 'thoreau/setup_assembler'
-require_relative 'thoreau/assertion_block'
-require_relative 'thoreau/equivalence_class'
-require_relative 'thoreau/dsl_context'
 require_relative 'thoreau/v2/dsl'
 
 module Thoreau
