+++ lib/thoreau/version.rb
@@ -1,3 +1,3 @@
 module Thoreau
-  VERSION = "0.1.0"
+  VERSION = "0.2.0"
 end
